src/requirements.txt
influxdb_client==1.27.0
python-dotenv==0.20.0
pyzmq==22.3.0
